refactor(wow-core): optimize wait mechanism and simplify waiting conditions
- Introduce WaitingForFunction and WaitingForStage abstract classes
- Update existing waiting classes to inherit from new abstract classes
- Simplify isWaitingForSignal logic in WaitingForFunction
- Remove redundant code and improve overall structure
