+++ wow-core/src/main/kotlin/me/ahoo/wow/command/wait/WaitingForAfterProcessed.kt
@@ -13,11 +13,9 @@
 
 package me.ahoo.wow.command.wait
 
-import me.ahoo.wow.api.messaging.function.FunctionNameCapable
-import me.ahoo.wow.api.messaging.processor.ProcessorInfo
 import reactor.core.publisher.Mono
 
-abstract class WaitingForAfterProcessed : AbstractWaitingFor(), ProcessorInfo, FunctionNameCapable {
+abstract class WaitingForAfterProcessed : AbstractWaitingFor() {
     @Volatile
     private var processedSignal: WaitSignal? = null
 
@@ -32,19 +30,7 @@ abstract class WaitingForAfterProcessed : AbstractWaitingFor(), ProcessorInfo, F
     }
 
     open fun isWaitingForSignal(signal: WaitSignal): Boolean {
-        if (signal.stage != stage || !isSameBoundedContext(signal.function)) {
-            return false
-        }
-        if (processorName.isBlank()) {
-            return true
-        }
-        if (processorName != signal.function.processorName) {
-            return false
-        }
-        if (functionName.isBlank()) {
-            return true
-        }
-        return signal.function.name == functionName
+        return signal.stage == stage
     }
 
     override fun waitingLast(): Mono<WaitSignal> {
