+++ wow-core/src/main/kotlin/me/ahoo/wow/command/wait/WaitingForEventHandled.kt
@@ -17,7 +17,7 @@ class WaitingForEventHandled(
     override val contextName: String,
     override val processorName: String = "",
     override val functionName: String = ""
-) : WaitingForAfterProcessed() {
+) : WaitingForFunction() {
     override val stage: CommandStage
         get() = CommandStage.EVENT_HANDLED
 }
